UPDATE: Warning messages and examples
Added warning messages for the "googledrive" permissions in the functions that donwload data from IEMA, IMAZON and SEEG.

Altered the README examples of COMEX, PEVS and TERRACLIMATE< reducing time spans or setting less especific geo_levels, in order to reduce data size. Also added a message informing that downloading might take some time if time periods are too big or if the level is municipality (for PEVS only)

Could not think of any better example regarding BACI, as the minimum time period that can be defined is ony year, which is already too much, and there is no argument for municipality. Only added a message warning the user about size and time.