READ.ME and SEEG update (#156)
* Update README.Rmd

* Update SEEG.Rmd

* UPDATE: Warning messages and examples

Added warning messages for the "googledrive" permissions in the functions that donwload data from IEMA, IMAZON and SEEG.

Altered the README examples of COMEX, PEVS and TERRACLIMATE< reducing time spans or setting less especific geo_levels, in order to reduce data size. Also added a message informing that downloading might take some time if time periods are too big or if the level is municipality (for PEVS only)

Could not think of any better example regarding BACI, as the minimum time period that can be defined is ony year, which is already too much, and there is no argument for municipality. Only added a message warning the user about size and time.

* Revert "UPDATE: Warning messages and examples"

This reverts commit b8804e9d5c2fb4932b2c7360107e41e5041f7d39.

* Update README.Rmd

* Update README.md

* GoogleDrive vignette and README.Rmd knit

Added a new vignette for googledrive but I didn't insert indo the README.Rmd yet. In addition, I tried to knit the README.Rmd and came across some problems, but now it seems to be working

* Message removed from SEEG.Rmd and added to GOOGLEDRIVE.Rmd

* adding SEEG vignette from Master

* editing the googledrive vignette

* adding googledrive vignette and warnings to readme

* adding red ball for warning

* adding vignette to automatic message

message when the googledrive package is called from within our funtions

* adding misc vignettes to readme table of contents

also rewording the googledrive readme a bit

Co-authored-by: AntonioBergallo <antoniorbergallo@gmail.com>
Co-authored-by: Igor Rigolon <igor.rilave@hotmail.com>
Co-authored-by: Antonio Bergallo <113378201+AntonioBergallo@users.noreply.github.com>
Co-authored-by: IgorRigolon <77173467+IgorRigolon@users.noreply.github.com>