load_epe expansion and improvement, load_ben removal
Now load_epe has two datasets:
- national_energy_balance (old database loaded by load_ben)
- energy_consumption_per_class (merged both datasets loaded previously by load_epe)

The last dataset is not ready. It doesn't load subsystems other than isolated system. It is also not in an optimal presenting form.

Some improvements on the script to keep it on datazoom.amazonia standards.

Added a default time_period that varies depending on the dataset.